requirements.txt
asgiref==3.4.1
certifi==2021.5.30
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.6
django-autoslug==1.9.8
django-cors-headers==3.8.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-phone-field==1.8.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
flake8==3.9.2
fsspec==2021.8.1
gunicorn==20.1.0
MarkupPy==1.14
mccabe==0.6.1
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.2
Pillow==8.3.1
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
